move connection logic to its own file
